pyproject.toml
[project]
name = "olx-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-ckeditor-5>=0.2.19",
    "django-jsonform>=2.23.2",
    "django-mptt>=0.18.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
